Switch from Callbacks to listeners in the display tab
Former-commit-id: 8cb2a4708ce78c1c5bb05c36b4e642a07cafb4c9 [formerly 59c8f3df7fc277f9a72449352ad6d826cc8595f8]
Former-commit-id: 348b2b581e675f1e9335015340db06f1128a74c4
Former-commit-id: 3decfd4ebf0e1e211771db4aedb975be2aa74537
Former-commit-id: 276df914e5c6a1411b9ec0d821ef4cc84e0b68d5 [formerly 12eb7d0729c38ef47018704aade33ed7d6e8c00f]
Former-commit-id: c4b76ab6ef9588bdd81fe666ddf50a6f5c12a19c
Former-commit-id: 1174ff6544ebf09e637c505c9cfcbce861ea2473

+++ classes/@transceiver_cl/get_pulse_length.m
@@ -1,6 +1,6 @@
 function [T,Np]=get_pulse_length(Trans_obj)
 
-Np=round(Trans_obj.Params.PulseLength(1)/Trans_obj.Params.SampleInterval(1));
-T=Np*(Trans_obj.Params.SampleInterval(1));
+Np=double(round(Trans_obj.Params.PulseLength(1)/Trans_obj.Params.SampleInterval(1)));
+T=double(Np*(Trans_obj.Params.SampleInterval(1)));
 
 end
